# api/core/openapi.yaml
openapi: 3.0.3
info:
  title: Flight Collaboration Platform Core API
  version: 1.0.0
  description: Internal management and Core business boundary. Core owns business truth.
servers:
  - url: /
paths:
...
  /internal/identity/v1/password/verify:
    post:
      summary: Verify an employee password for the Edge identity boundary
      description: Internal Core-owned contract. Requires X-Internal-Identity-Key in local shared-key mode.
      x-internal: true
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PasswordVerifyRequest'}
      responses:
        '200':
          description: Password verified or binding required
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PasswordVerifyResponse'}
        '401': {description: Invalid credentials or internal identity key}
        '403': {description: Staff is inactive}
  /internal/identity/v1/staff/status:
    post:
      summary: Resolve the current active status of a Core Staff
      x-internal: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [public_id]
              properties:
                public_id: {type: string, format: uuid}
      responses:
        '200': {description: Active Staff returned}
        '403': {description: Staff is inactive}
  /internal/identity/v1/providers/exchange:
    post:
      summary: Resolve a provider authorization code to a Core Staff
      x-internal: true
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ProviderExchangeRequest'}
      responses:
        '200': {description: Staff identity resolved}
        '403': {description: Provider identity is unmapped or staff is inactive}
  /internal/identity/v1/bindings/complete:
    post:
      summary: Complete a Core-owned one-time external identity binding
      x-internal: true
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/BindingCompleteRequest'}
      responses:
        '200': {description: Binding completed and Staff returned}
        '401': {description: Invalid provider code or binding ticket}
        '409': {description: External identity binding conflict}
...
  /api/v1/personnel/status:
    get:
      summary: List current personnel work states
      security: [{bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/PageSize'}
        - {name: work_state, in: query, schema: {type: string}}
        - {name: team_public_id, in: query, schema: {type: string, format: uuid}}
        - {name: area_public_id, in: query, schema: {type: string, format: uuid}}
      responses:
        '200':
          description: Paginated personnel status
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PersonnelStatusListResponse'}
  /api/v1/personnel/{publicID}/status-history:
    get:
      summary: List one personnel status history
      security: [{bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/PublicID'}
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/PageSize'}
      responses:
        '200':
          description: Paginated status history
          content:
            application/json:
              schema: {$ref: '#/components/schemas/StatusHistoryListResponse'}
  /api/v1/events:
    get:
      summary: List operational events and delivery state
      security: [{bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/PageSize'}
        - {name: event_type, in: query, schema: {type: string}}
        - {name: status, in: query, schema: {type: string}}
        - {name: flight_public_id, in: query, schema: {type: string, format: uuid}}
        - {name: from, in: query, schema: {type: string, format: date-time}}
        - {name: to, in: query, schema: {type: string, format: date-time}}
      responses:
        '200':
          description: Paginated event stream read model
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EventListResponse'}
...
  /api/v1/audit:
    get:
      summary: List redacted audit records
      security: [{bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/PageSize'}
        - {name: actor_id, in: query, schema: {type: string}}
        - {name: action, in: query, schema: {type: string}}
        - {name: resource_type, in: query, schema: {type: string}}
        - {name: from, in: query, schema: {type: string, format: date-time}}
        - {name: to, in: query, schema: {type: string, format: date-time}}
      responses:
        '200':
          description: Paginated audit records
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuditListResponse'}
  /api/v1/scopes:
    get:
      summary: Return the current Core-derived scope
      security: [{bearerAuth: []}]
      responses:
        '200':
          description: Principal roles and scope
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ScopeResponse'}
  /api/v1/diagnostics:
    get:
      summary: Return safe Core runtime and synchronization diagnostics
      security: [{bearerAuth: []}]
      responses:
        '200':
          description: Diagnostics without credentials or raw payloads
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DiagnosticsResponse'}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    flightSourceKey:
      type: apiKey
      in: header
      name: X-Flight-Source-Key
  parameters:
    Page:
      name: page
      in: query
      schema: {type: integer, minimum: 1, default: 1}
    PageSize:
      name: page_size
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
...
    PublicID:
      name: publicID
      in: path
      required: true
      schema: {type: string, format: uuid}
  schemas:
...
    PasswordVerifyRequest:
      type: object
      required: [employee_no, password, client]
      properties:
        employee_no: {type: string}
        password: {type: string, format: password}
        client: {type: string, enum: [employee-miniapp, employee-wecom-miniapp, employee-web]}
        provider: {type: string, enum: [personal_wechat, wecom]}
        provider_app: {type: string}
    PasswordVerifyResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [state, staff]
          properties:
            state: {type: string, enum: [authenticated, binding_required]}
            binding_ticket: {type: string, format: uuid}
            staff: {$ref: '#/components/schemas/StaffIdentity'}
    ProviderExchangeRequest:
      type: object
      required: [provider, provider_code, client]
      properties:
        provider: {type: string, enum: [personal_wechat, wecom]}
        provider_code: {type: string, format: password}
        client: {type: string, enum: [employee-miniapp, employee-wecom-miniapp, employee-web]}
        redirect_uri: {type: string, format: uri}
    BindingCompleteRequest:
      type: object
      required: [binding_ticket, provider, provider_code, client]
      properties:
        binding_ticket: {type: string, format: uuid}
        provider: {type: string, enum: [personal_wechat, wecom]}
        provider_code: {type: string, format: password}
        client: {type: string, enum: [employee-miniapp, employee-wecom-miniapp, employee-web]}
        redirect_uri: {type: string, format: uri}
    StaffIdentity:
      type: object
      required: [public_id, employee_no, display_name, role]
      properties:
        public_id: {type: string, format: uuid}
        employee_no: {type: string}
        display_name: {type: string}
        role: {type: string}
...
    PersonnelStatus:
      type: object
      required: [public_id, employee_no, display_name, area_public_id, area_name, team_public_id, team_name, position_code, capability_code, work_state, status_version, last_state_changed_at, enabled]
      properties:
        public_id: {type: string, format: uuid}
        employee_no: {type: string}
        display_name: {type: string}
        area_public_id: {type: string, format: uuid}
        area_name: {type: string}
        team_public_id: {type: string, format: uuid}
        team_name: {type: string}
        position_code: {type: string}
        capability_code: {type: string}
        work_state: {type: string, enum: [idle, reserved, busy, unavailable]}
        status_version: {type: integer, format: int64, minimum: 0}
        last_state_changed_at: {type: string, format: date-time}
        unavailable_reason: {type: string}
        enabled: {type: boolean}
    StatusHistory:
      type: object
      required: [public_id, personnel_public_id, status_version, to_state, actor_type, occurred_at]
      properties:
        public_id: {type: string, format: uuid}
        personnel_public_id: {type: string, format: uuid}
        status_version: {type: integer, format: int64, minimum: 0}
        from_state: {type: string}
        to_state: {type: string}
        reason: {type: string}
        actor_type: {type: string}
        actor_public_id: {type: string, format: uuid}
        assignment_id: {type: integer, format: int64}
        command_id: {type: string}
        occurred_at: {type: string, format: date-time}
    Event:
      type: object
      required: [public_id, event_type, status, aggregate_type, aggregate_public_id, source, occurred_at]
      properties:
        public_id: {type: string, format: uuid}
        event_type: {type: string}
        status: {type: string}
        aggregate_type: {type: string}
        aggregate_public_id: {type: string, format: uuid}
        flight_public_id: {type: string, format: uuid}
        flight_display_no: {type: string}
        source: {type: string}
        occurred_at: {type: string, format: date-time}
        last_error: {type: string}
    AuditEntry:
      type: object
      required: [id, actor_type, actor_id, action, resource_type, resource_id, result, request_id, trace_id, source_ip, occurred_at]
      properties:
        id: {type: integer, format: int64}
        actor_type: {type: string}
        actor_id: {type: string}
        action: {type: string}
        resource_type: {type: string}
        resource_id: {type: string}
        result: {type: string}
        request_id: {type: string}
        trace_id: {type: string}
        source_ip: {type: string}
        occurred_at: {type: string, format: date-time}
    PersonnelStatusListResponse:
      type: object
      required: [data, request_id, trace_id]
      properties:
        data:
          type: object
          required: [items, page, page_size, total]
          properties:
            items: {type: array, items: {$ref: '#/components/schemas/PersonnelStatus'}}
            page: {type: integer, minimum: 1}
            page_size: {type: integer, minimum: 1, maximum: 100}
            total: {type: integer, format: int64, minimum: 0}
        request_id: {type: string}
        trace_id: {type: string}
    StatusHistoryListResponse:
      type: object
      required: [data, request_id, trace_id]
      properties:
        data:
          type: object
          required: [items, page, page_size, total]
          properties:
            items: {type: array, items: {$ref: '#/components/schemas/StatusHistory'}}
            page: {type: integer, minimum: 1}
            page_size: {type: integer, minimum: 1, maximum: 100}
            total: {type: integer, format: int64, minimum: 0}
        request_id: {type: string}
        trace_id: {type: string}
    EventListResponse:
      type: object
      required: [data, request_id, trace_id]
      properties:
        data:
          type: object
          required: [items, page, page_size, total]
          properties:
            items: {type: array, items: {$ref: '#/components/schemas/Event'}}
            page: {type: integer, minimum: 1}
            page_size: {type: integer, minimum: 1, maximum: 100}
            total: {type: integer, format: int64, minimum: 0}
        request_id: {type: string}
        trace_id: {type: string}
    AuditListResponse:
      type: object
      required: [data, request_id, trace_id]
      properties:
        data:
          type: object
          required: [items, page, page_size, total]
          properties:
            items: {type: array, items: {$ref: '#/components/schemas/AuditEntry'}}
            page: {type: integer, minimum: 1}
            page_size: {type: integer, minimum: 1, maximum: 100}
            total: {type: integer, format: int64, minimum: 0}
        request_id: {type: string}
        trace_id: {type: string}
    ScopeResponse:
      type: object
      required: [data, request_id, trace_id]
      properties:
        data:
          type: object
          required: [principal_public_id, roles, global, area_ids, team_ids]
          properties:
            principal_public_id: {type: string, format: uuid}
            roles: {type: array, items: {type: string}}
            global: {type: boolean}
            area_ids: {type: array, items: {type: integer, format: int64}}
            team_ids: {type: array, items: {type: integer, format: int64}}
        request_id: {type: string}
        trace_id: {type: string}
    DiagnosticsResponse:
      type: object
      required: [data, request_id, trace_id]
      properties:
        data:
          type: object
          required: [component, generated_at, database_reachable, runtime, sync]
          properties:
            component: {type: string}
            generated_at: {type: string, format: date-time}
            database_reachable: {type: boolean}
            runtime:
              type: object
              required: [environment, redis_enabled, flight_source_configured, real_identity_provider, development_actor_headers]
              properties:
                environment: {type: string}
                redis_enabled: {type: boolean}
                flight_source_configured: {type: boolean}
                real_identity_provider: {type: boolean}
                development_actor_headers: {type: boolean}
            sync:
              type: object
              required: [flight_source_pending, flight_source_retry, flight_source_failed, outbox_pending, outbox_failed, core_inbox_failed]
              properties:
                flight_source_pending: {type: integer, format: int64, minimum: 0}
                flight_source_retry: {type: integer, format: int64, minimum: 0}
                flight_source_failed: {type: integer, format: int64, minimum: 0}
                outbox_pending: {type: integer, format: int64, minimum: 0}
                outbox_failed: {type: integer, format: int64, minimum: 0}
                core_inbox_failed: {type: integer, format: int64, minimum: 0}
        request_id: {type: string}
        trace_id: {type: string}
...